requirements.txt
aiohttp==3.9.1
aiosignal==1.3.2
aiosqlite==0.19.0
alembic==1.13.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==25.1.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
cryptography==41.0.7
exceptiongroup==1.2.2
Flask==3.0.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.25.2
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.1
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
python-i18n==0.3.9
python-socks==2.4.3
python-telegram-bot==20.7
sniffio==1.3.1
socksio==1.0.0
SQLAlchemy==2.0.23
typing_extensions==4.12.2
Werkzeug==3.0.1
yarl==1.18.3
zipp==3.21.0
